requirements.txt
fastapi==0.111.0
motor==3.4.0
pydantic==2.7.1
pymongo==4.7.2
python-dotenv==1.0.1
